サーバーレス用ににディレクトリ構成変更
