+++ vercel.json
@@ -1,16 +1,15 @@
 {
     "version": 2,
     "builds": [
-      {
-        "src": "server.js",
-        "use": "@vercel/node"
-      }
+        {
+            "src": "api/notes.js",
+            "use": "@vercel/node"
+        }
     ],
     "routes": [
-      {
-        "src": "/(.*)",
-        "dest": "server.js"
-      }
+        {
+            "src": "/api/notes",
+            "dest": "/api/notes.js"
+        }
     ]
-  }
-  
+}
